fix(types): move backend types
