+++ src/services/flag/types.ts
@@ -1,0 +1,6 @@
+import { UUID } from 'src/types';
+
+export type Flag = {
+  id: UUID;
+  name: string;
+};
