+++ src/services/index.ts
@@ -1,9 +1,14 @@
 export * from './app';
 export * from './database';
 export * from './file';
+export * from './flag';
 export * from './h5p';
 export * from './etherpad';
 export * from './item-memberships';
+export * from './item-tags';
 export * from './items';
+export * from './invitation';
 export * from './members';
 export * from './public';
+export * from './chat';
+export * from './categories';
