+++ src/services/item-tags/index.ts
@@ -1,0 +1,1 @@
+export * from './types';
