+++ src/services/item-tags/types.ts
@@ -1,0 +1,9 @@
+import { UUID } from 'src/types';
+
+export type ItemTag = {
+  id: UUID;
+  itemPath: string;
+  tagId: UUID;
+  createdAt: string;
+  creator: string;
+};
