+++ src/types.ts
@@ -1,0 +1,2 @@
+// alias type for uuid v4
+export type UUID = string;
